+++ crates/compiler/src/driver_constants.rs
@@ -210,7 +210,7 @@ pub const ECHO_BUFFER_MAX_EDL: u8 = 15;
 // Song constants
 pub const MAX_SUBROUTINES: usize = 255;
 
-pub const SONG_HEADER_SIZE: usize = N_MUSIC_CHANNELS * 2 + 13;
+pub const SONG_HEADER_SIZE: usize = N_MUSIC_CHANNELS * 2 + 14;
 pub const SONG_HEADER_CHANNELS_SIZE: usize = N_MUSIC_CHANNELS * 2;
 pub const SONG_HEADER_TICK_TIMER_OFFSET: usize = SONG_HEADER_SIZE - 2;
 pub const SONG_HEADER_N_SUBROUTINES_OFFSET: usize = SONG_HEADER_SIZE - 1;